GDRCD 5.6.0.4 - Aggiornamento controlli Gestione > Manutenzione (#262)
* GDRCD [UPD] Aggiunto metodo per il controllo permessi utente

* GESTIONE [UPD] Allineamento gestione manutenzione con il nuovo modulo di caricamento delle pagine e puliti i sorgenti; correzione nel controllo permessi

* GESTIONE [FIX] Spostato controllo permessi utente prima della creazione del contenuto della pagina di Gestione > Manutenzione

+++ pages/gestione/manutenzione/old_messages.inc.php
@@ -1,4 +1,12 @@
 <?php
+
+/*HELP: */
+/*Controllo permessi utente*/
+if(!gdrcd_controllo_permessi($PARAMETERS['administration']['maintenance']['access_level'])) {
+    echo '<div class="error">'.gdrcd_filter('out', $MESSAGE['error']['not_allowed']).'</div>';
+    die();
+}
+
 if((is_numeric($_POST['mesi']) === true) && ($_POST['mesi'] >= 0) && ($_POST['mesi'] <= 12)) {
     /*Eseguo l'aggiornamento*/
     gdrcd_query("DELETE FROM messaggi WHERE DATE_SUB(NOW(), INTERVAL ".gdrcd_filter('num', $_POST['mesi'])." MONTH) > spedito");
@@ -7,7 +15,7 @@ if((is_numeric($_POST['mesi']) === true) && ($_POST['mesi'] >= 0) && ($_POST['me
     gdrcd_query("OPTIMIZE TABLE backmessaggi");
     ?>
     <!-- Conferma -->
-    <div class="warning">
+    <div class="success">
         <?php echo gdrcd_filter('out', $MESSAGE['warning']['modified']); ?>
     </div>
     <?php
@@ -18,10 +26,3 @@ if((is_numeric($_POST['mesi']) === true) && ($_POST['mesi'] >= 0) && ($_POST['me
     </div>
     <?php
 }
-?>
-<!-- Link di ritorno alla visualizzazione di base -->
-<div class="link_back">
-    <a href="main.php?page=gestione_manutenzione">
-        <?php echo gdrcd_filter('out', $MESSAGE['interface']['administration']['maintenance']['link']['back']); ?>
-    </a>
-</div>
